+++ src/kbucket/bucket.rs
@@ -56,7 +56,7 @@ pub enum NodeInsertError<TNode> {
     MismatchVersion(TNode, Version),
 }
 
-impl<'a, TNode> NodeInsertOk<'a, TNode> {
+impl<TNode> NodeInsertOk<'_, TNode> {
     /// Returns an optional reference to the node pending eviction.
     pub fn pending_eviction(&self) -> Option<&TNode> {
         match self {
@@ -325,7 +325,25 @@ mod tests {
 
     #[test]
     fn test_lru_base_5secs() -> Result<()> {
+        // Create all the nodes at the beginning to ensure that PoW is not a
+        // factor in the test timing.
         let root = PeerNode::generate("127.0.0.1:666", 0)?;
+        let node1 = PeerNode::generate("192.168.1.1:8080", 0)?;
+        let node1_copy = PeerNode::generate("192.168.1.1:8080", 0)?;
+        let node1_copy2 = PeerNode::generate("192.168.1.1:8080", 0)?;
+        let node2 = PeerNode::generate("192.168.1.2:8080", 0)?;
+
+        let mut additionals = vec![];
+        for i in 2..21 {
+            additionals.push(PeerNode::generate(
+                &format!("192.168.1.{}:8080", i)[..],
+                0,
+            )?);
+        }
+
+        let pending = PeerNode::generate("192.168.1.21:8080", 0)?;
+        let pending_2 = PeerNode::generate("192.168.1.21:8080", 0)?;
+
         let mut config = BucketConfig::default();
         config.node_evict_after = Duration::from_millis(1000);
         config.node_ttl = Duration::from_secs(5);
@@ -333,9 +351,7 @@ mod tests {
         let mut route_table = Tree::new(root, config);
 
         let bucket = route_table.bucket_for_test();
-        let node1 = PeerNode::generate("192.168.1.1:8080", 0)?;
         let id_node1 = node1.id().as_binary().clone();
-        let node1_copy = PeerNode::generate("192.168.1.1:8080", 0)?;
         match bucket.insert(node1).expect("This should return an ok()") {
             NodeInsertOk::Inserted { .. } => {}
             _ => assert!(false),
@@ -351,7 +367,6 @@ mod tests {
             _ => assert!(false),
         }
         assert_eq!(Some(&id_node1), bucket.last_id());
-        let node2 = PeerNode::generate("192.168.1.2:8080", 0)?;
         let id_node2 = node2.id().as_binary().clone();
 
         match bucket.insert(node2).expect("This should return an ok()") {
@@ -364,7 +379,7 @@ mod tests {
         assert_eq!(Some(&id_node1), bucket.least_used_id());
 
         match bucket
-            .insert(PeerNode::generate("192.168.1.1:8080", 0)?)
+            .insert(node1_copy2)
             .expect("This should return an ok()")
         {
             NodeInsertOk::Updated { .. } => {}
@@ -378,14 +393,8 @@ mod tests {
         assert_eq!(&id_node2, a.unwrap().id().as_binary());
         assert_eq!(Some(&id_node1), bucket.last_id());
         assert_eq!(Some(&id_node1), bucket.least_used_id());
-        for i in 2..21 {
-            match bucket
-                .insert(PeerNode::generate(
-                    &format!("192.168.1.{}:8080", i)[..],
-                    0,
-                )?)
-                .expect("This should return an ok()")
-            {
+        for n in additionals {
+            match bucket.insert(n).expect("This should return an ok()") {
                 NodeInsertOk::Inserted { .. } => {
                     assert!(bucket.pick::<K_BETA>().count() <= K_BETA);
                 }
@@ -393,7 +402,6 @@ mod tests {
             }
         }
         assert_eq!(bucket.pick::<K_BETA>().count(), K_BETA);
-        let pending = PeerNode::generate("192.168.1.21:8080", 0)?;
         let pending_id = pending.id().as_binary().clone();
         match bucket.insert(pending).expect_err("this should be error") {
             NodeInsertError::Full(pending) => {
@@ -409,9 +417,9 @@ mod tests {
                             &pending_id
                         );
                         thread::sleep(Duration::from_secs(1));
-                        let pending =
-                            PeerNode::generate("192.168.1.21:8080", 0)?;
-                        match bucket.insert(pending).expect("this should be ok")
+                        match bucket
+                            .insert(pending_2)
+                            .expect("this should be ok")
                         {
                             NodeInsertOk::Inserted { inserted: _ } => {}
                             v => {